requirements.txt
google-api-python-client==1.12.8
google-cloud-bigquery==2.6.0
pandas==1.1.5
pyarrow==2.0.0